requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
frozenlist==1.7.0
idna==3.10
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
multidict==6.6.4
narwhals==2.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
rich==13.7.0
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
